server/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
langchain==0.1.20
langchain-openai==0.1.1
langchain-community==0.0.38
chromadb==0.4.22
pypdf==3.17.4
python-docx==1.1.0
openai==1.14.0
Werkzeug==3.0.1
